requirements_test.txt
-r requirements_dev.txt
asynctest >= 0.13
pytest >= 7.2.0
pytest-cov >= 2.10
pytest-homeassistant-custom-component>=0.13.18
pytest-aiohttp >= 0.3.0
aiodiscover >= 1.4.0
scapy >= 2.5.0